Version 1.2.1 - fix for Nunit Testing.

+++ Scorchio.VisualStudio/Services/TraceService.cs
@@ -8,8 +8,7 @@ namespace Scorchio.VisualStudio.Services
     using System;
     using System.Diagnostics;
     using System.IO;
-
-    using Microsoft.Win32;
+    using System.Windows.Forms;
 
     /// <summary>
     ///  Defines the TraceService type.
@@ -17,16 +16,39 @@ namespace Scorchio.VisualStudio.Services
     public static class TraceService
     {
         /// <summary>
-        /// Gets a value indicating whether [log to file].
+        /// The log to trace setting.
         /// </summary>
+        private static bool logToTrace = true;
+
+        /// <summary>
+        /// The log to file setting.
+        /// </summary>
+        private static bool logToFile;
+
+        /// <summary>
+        /// The log file setting.
+        /// </summary>
+        private static string logFile;
+
+        /// <summary>
+        /// The display errors setting.
+        /// </summary>
+        private static bool displayErrors;
+
+        /// <summary>
+        /// Gets a value indicating whether to log to trace.
+        ///  </summary>
+        internal static bool LogToTrace
+        {
+            get { return logToTrace; }
+        }
+
+        /// <summary>
+        /// Gets a value indicating whether to log to file.
+        ///  </summary>
         internal static bool LogToFile
         {
-            get 
-            {
-                string value = (string)Registry.CurrentUser.GetValue(@"SOFTWARE\Scorchio Limited\Ninja Coder for MvvmCross\LogToFile", "N");
-
-                return value == "Y";
-            }
+            get { return logToFile; }
         }
 
         /// <summary>
@@ -34,10 +56,34 @@ namespace Scorchio.VisualStudio.Services
         /// </summary>
         internal static string LogFile
         {
-            get
-            {
-                return (string)Registry.CurrentUser.GetValue(@"SOFTWARE\Scorchio Limited\Ninja Coder for MvvmCross\LogFilePath", Environment.GetFolderPath(Environment.SpecialFolder.DesktopDirectory) + "\\ninja-coder-for-mvvmcross.log");
-            }
+            get { return logFile; }
+        }
+
+        /// <summary>
+        /// Gets a value indicating whether to display Errors.
+        /// </summary>
+        internal static bool DisplayErrors
+        {
+            get { return displayErrors; }
+        }
+
+        /// <summary>
+        /// Initializes the specified settings.
+        /// </summary>
+        /// <param name="logToTraceSetting">if set to <c>true</c> [log to trace setting].</param>
+        /// <param name="logToFileSetting">if set to <c>true</c> [log to file setting].</param>
+        /// <param name="logFileSetting">The log file setting.</param>
+        /// <param name="displayErrorsSetting">if set to <c>true</c> [display errors setting].</param>
+        public static void Initialize(
+            bool logToTraceSetting,
+            bool logToFileSetting,
+            string logFileSetting,
+            bool displayErrorsSetting)
+        {
+            logToTrace = logToTraceSetting;
+            logToFile = logToFileSetting;
+            logToFile = logToFileSetting;
+            displayErrors = displayErrorsSetting;
         }
 
         /// <summary>
@@ -46,11 +92,14 @@ namespace Scorchio.VisualStudio.Services
         /// <param name="message">The message.</param>
         public static void WriteLine(string message)
         {
-            Trace.WriteLine(GetTimedMessage(message));
+            if (LogToTrace)
+            {
+                Trace.WriteLine(GetTimedMessage(string.Empty, message));
+            }
 
             if (LogToFile)
             {
-                WriteMessageToLogFile(GetTimedMessage(message));
+                WriteMessageToLogFile(GetTimedMessage(string.Empty, message));
             }
         }
 
@@ -60,13 +109,21 @@ namespace Scorchio.VisualStudio.Services
         /// <param name="message">The message.</param>
         public static void WriteError(string message)
         {
-            string timedMessage = "**ERROR** " + GetTimedMessage(message);
+            string timedMessage = GetTimedMessage("**ERROR**", message);
 
-            Trace.WriteLine(timedMessage);
+            if (LogToTrace)
+            {
+                Trace.WriteLine(timedMessage);
+            }
 
             if (LogToFile)
             {
                 WriteMessageToLogFile(timedMessage);
+            }
+
+            if (DisplayErrors)
+            {
+                MessageBox.Show(message, "Ninja Coder for MvvmCross");
             }
         }
 
@@ -76,19 +133,27 @@ namespace Scorchio.VisualStudio.Services
         /// <param name="message">The message.</param>
         internal static void WriteMessageToLogFile(string message)
         {
-            StreamWriter sw = new StreamWriter(LogFile, true);
-            sw.WriteLine(message);
-            sw.Close();
+            if (File.Exists(LogFile))
+            {
+                StreamWriter sw = new StreamWriter(LogFile, true);
+                sw.WriteLine(message);
+                sw.Close();
+            }
         }
 
         /// <summary>
         /// Gets the timed message.
         /// </summary>
+        /// <param name="type">The type.</param>
         /// <param name="message">The message.</param>
-        /// <returns>A timed stamped message</returns>
-        internal static string GetTimedMessage(string message)
+        /// <returns>
+        /// A timed stamped message
+        /// </returns>
+        internal static string GetTimedMessage(
+            string type,
+            string message)
         {
-            return string.Format("{0} {1}", DateTime.Now.ToString("dd MMM yy HH:mm:ss fff"), message);
+            return string.Format("{0} {1} {2}", DateTime.Now.ToString("dd MMM yy HH:mm:ss fff"), type, message);
         }
    }
 }
